requirements.txt
fastapi
uvicorn
pytest


Flask==2.1.0
openai==0.27.0
pytest==7.0.1
python-dotenv==0.19.2
